ducklake/pyproject.toml
[project]
name = "ducklake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka>=2.11.1",
    "duckdb>=1.4.1",
    "numpy>=2.3.3",
    "pyspark>=4.0.1",
    "python-dotenv>=1.1.1",
]
